move the process_data method to the Data parent class
