+++ swan/dataset/graph/molecular_graph.py
@@ -4,26 +4,26 @@ Index
 -----
 .. currentmodule:: swan.graph.molecular_graph
 .. autosummary::
-    create_molecular_graph_data
+    create_molecular_torch_geometric_graph
 
 
 API
 ---
-.. autofunction:: create_molecular_graph_data
+.. autofunction:: create_molecular_torch_geometric_graph
+
 """
 import numpy as np
 import torch
 from rdkit import Chem
 from torch import Tensor
-from torch_geometric.data import Data
+import torch_geometric as tg
 
 from swan.dataset.features.featurizer import (compute_molecular_graph_edges,
                                               generate_molecular_features)
 
 
-def create_molecular_graph_data(mol: Chem.rdchem.Mol,
-                                positions: Tensor = None,
-                                labels: Tensor = None) -> Data:
+def create_molecular_torch_geometric_graph(
+        mol: Chem.rdchem.Mol, positions: Tensor = None, labels: Tensor = None) -> tg.data.Data:
     """Create a torch-geometry data object representing a graph.
 
     See torch-geometry documentation:
@@ -37,8 +37,9 @@ def create_molecular_graph_data(mol: Chem.rdchem.Mol,
     atomic_features = torch.tensor(atomic_features.astype(np.float32))
     bond_features = torch.tensor(bond_features.astype(np.float32))
 
-    return Data(x=atomic_features,
-                edge_attr=bond_features,
-                edge_index=edges,
-                positions=positions,
-                y=labels)
+    return tg.data.Data(
+        x=atomic_features,
+        edge_attr=bond_features,
+        edge_index=edges,
+        positions=positions,
+        y=labels)
